1.添加世界坐标系 2.优化HDR 3.优化数学库 3.优化材质更新

+++ include/ray/game_application.h
@@ -47,7 +47,7 @@ public:
 	GameApplication() noexcept;
 	virtual ~GameApplication() noexcept;
 
-	bool open(WindHandle hwnd, std::size_t width, std::size_t height) noexcept;
+	bool open(WindHandle hwnd, std::uint32_t width, std::uint32_t height) noexcept;
 	void close() noexcept;
 
 	bool start() noexcept;
